+++ packages/core/tests/ExecuteScriptRoutine.test.ts
@@ -91,7 +91,7 @@ describe('ExecuteScriptRoutine', () => {
         routine.context.args.workspaces = '*';
         routine.context.workspaces = ['packages/*'];
         routine.context.workspaceRoot = fixturePath;
-        routine.context.root = fixturePath;
+        routine.context.cwd = fixturePath;
       });
 
       it('adds a routine for each workspace', () => {